move endpoints health check controller to separate package
