+++ pkg/libs/endpointaccessible/endpoint_accessible_controller_test.go
@@ -3,7 +3,6 @@ package endpointaccessible
 import (
 	"context"
 	"fmt"
-	"reflect"
 	"testing"
 
 	"github.com/openshift/library-go/pkg/controller/factory"
@@ -51,27 +50,6 @@ func Test_endpointAccessibleController_sync(t *testing.T) {
 			}
 			if err := c.sync(context.Background(), factory.NewSyncContext(tt.name, events.NewInMemoryRecorder(tt.name))); (err != nil) != tt.wantErr {
 				t.Errorf("sync() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func Test_toHealthzURL(t *testing.T) {
-	tests := []struct {
-		name string
-		args []string
-		want []string
-	}{
-		{
-			name: "test urls",
-			args: []string{"a", "b"},
-			want: []string{"https://a/healthz", "https://b/healthz"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := toHealthzURL(tt.args); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("toHealthzURL() = %v, want %v", got, tt.want)
 			}
 		})
 	}
